Add structure for MECA
