+++ src/domain/submission/services/exporter/types.ts
@@ -1,0 +1,5 @@
+import { SubmissionId } from '../../types';
+
+export interface SubmissionExporter {
+    export(id: SubmissionId): Promise<Buffer>;
+}
