+++ src/domain/submission/services/storage/sftp-store.ts
@@ -1,4 +1,5 @@
-import { SubmissionWriter, PackageLocation, SubmissionId } from '../types';
+import { SubmissionId } from '../../types';
+import { PackageLocation, SubmissionWriter } from './types';
 
 export class SftpStore implements SubmissionWriter {
     private remotePath = 'somewhere-in-AWS';
